+++ tests/testthat/test_Filter.R
@@ -47,7 +47,7 @@ test_that("FilterForceLevel and FilterInject work inside function", {
     lg$debug("a debug message")
     lg$error("an error")
   }
-  expect_output(analyse(), "INFO.*debug.*type:.*INFO.*error.*type:.*")
+  expect_output(analyse(), "INFO.*debug.*type.*INFO.*error.*type.*")
   expect_length(lgr$filters, 0)
   lg$config(logger_config())
 })