+++ NEWS.md
@@ -1,11 +1,15 @@
-# lgr dev
+# lgr 0.4.3
 
 * `logger_index()` returns a `data.frame` with metadata on all registered
   loggers (#47) (thanks @Fuco1)
+  
 * export new `string_repr()` generic that is used to layout R objects for 
   formatted log message output (#48, thanks @mmuurr)
+  
 * The `$log()` method of Logger and LoggerGlue now unpacks conditons (except
   if they are supplied as a named argument) (#45, thanks @mmuurr)
+  
+* Fix some timezone related tests for CRAN
 
 
 # lgr 0.4.2